+++ packages/plugin-console-breadcrumbs/console-breadcrumbs.js
@@ -4,6 +4,12 @@ const { map, reduce, filter } = require('@bugsnag/core/lib/es-utils')
  * Leaves breadcrumbs when console log methods are called
  */
 exports.init = (client) => {
+  const isDev = /^dev(elopment)?$/.test(client.config.releaseStage)
+
+  const explicitlyDisabled = client.config.consoleBreadcrumbsEnabled === false
+  const implicitlyDisabled = (client.config.autoBreadcrumbs === false || isDev) && client.config.consoleBreadcrumbsEnabled !== true
+  if (explicitlyDisabled || implicitlyDisabled) return
+
   map(CONSOLE_LOG_METHODS, method => {
     const original = console[method]
     console[method] = (...args) => {