+++ CHANGELOG.md
@@ -6,6 +6,10 @@
 
 - Support serialising error objects (via. @bugsnag/safe-json-stringify@v4.0.0) (#356, #458)
 
+### Fixed
+
+- Fixed granular breadcrumb config logic (#461, #465, #466)
+
 ## 5.1.0 (2018-12-19)
 
 ### Fixed